+++ face/infra/base-bio-sdk/src/main/java/com/simprints/face/infra/basebiosdk/detection/Face.kt
@@ -23,7 +23,7 @@ data class Face(
     val yaw: Float,
     var roll: Float,
     val quality: Float,
-    val template: ByteArray,
+    val template: FloatArray,
     val format: String,
 ) {
     // Relative = coordinates are fractions of the source image dimensions
