+++ face/infra/base-bio-sdk/src/main/java/com/simprints/face/infra/basebiosdk/matching/FaceSample.kt
@@ -6,5 +6,5 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class FaceSample(
     val faceId: String,
-    val template: ByteArray,
+    val template: FloatArray,
 ) : Parcelable
