+++ fingerprint/infra/base-bio-sdk/src/main/java/com/simprints/fingerprint/infra/basebiosdk/matching/domain/Fingerprint.kt
@@ -2,6 +2,6 @@ package com.simprints.fingerprint.infra.basebiosdk.matching.domain
 
 class Fingerprint(
     val fingerId: FingerIdentifier,
-    val template: ByteArray,
+    val template: FloatArray,
     val format: String,
 )
